fix: add main to package.json for tsserver's benefit
This prevents a spurious tsserver 2307 error in VS Code and neovim.

This seems like a bug in tsserver, since tsc and ts-node can handle the
import just fine, but without a 'main', apparently it cannot resolve the
module properly, even though a 'main' field should not be strictly
necessary when using a moduleResolution setting of nodenext or node16.